[ui][code] Implement interactive signal adjustment
This patch adds an ability to interactively adjust parameters
of the generated signals. Once applied the following will be
available for the users:
  - Frequency can be adjusted while the signal is generated
  - Phase shift can be adjusted while the signal is generatted
  - Balance can be adjusted while signal is generated.

Phase shift is adjustable under the condition, that both channels
are running the same frequency.

On first press of the "Start" button within one session the generator
will warn the user about potentially high sound loud.